Add scripts to upload to GitHub from a third-party
This commit adds a shell script and README.md to show how we can perform
a lightweight mirroring of repositories not hosted in GitHub at analysis
time.

The analysis uses the "source archive" - the copy of the files analyzed
during the build - to populate a git repository, which is then pushed to
GitHub. The analyzed results are then submitted to GHAS on the same
repository.